RPMSG fixes
The important fix was the interrupt numbers in main.c around line 35,
see the comment immediately above.

Changing FW_RSC_ADDR_ANY from 0 to (~0) was probably important, but I
found that before I fixed the interrupt, so I'm not sure how big a deal
it was.

The other changes (message size and the constants at the top of the
resource deinition) are things I noticed were different in the example
code, but I don't know how important they are.